Refactor CPU-OS Communication and Enhance Error Handling
- Updated common.h to introduce CpuEvent enum for better clarity in CPU-OS communication events.
- Modified CPU class to replace direct memory writes for syscall results with a setCpuEvent method.
- Enhanced error handling in instruction parsing to provide more informative messages, including line numbers.
- Improved memory access checks in CPU to use defined constants for better readability.
- Added new fault handler addresses in common.h for arithmetic and unknown instruction faults.
- Updated main.cpp to support a new debug mode for detailed thread table dumps during context switches.
- Refactored memory loading functions to count lines read for better error reporting.
- Adjusted assembler tool to ensure argument count checks are type-safe.

+++ src/instruction.h
@@ -52,7 +52,7 @@ inline std::string opCodeToString(OpCode op)
         "USER",
         "SYSCALL_PRN", "SYSCALL_HLT_THREAD", "SYSCALL_YIELD",
         "UNKNOWN"};
-    if (static_cast<int>(op) < opCodeStrings.size())
+    if (static_cast<size_t>(op) < opCodeStrings.size())
     {
         return opCodeStrings[static_cast<int>(op)];
     }
